feature - client now sends the price as a parameter to the webservice socket

+++ holborn-car-park-client/src/main/java/uk/co/holborn/carparkclient/controllers/TicketDetailsPopUpController.java
@@ -33,15 +33,15 @@ public class TicketDetailsPopUpController implements Initializable {
         SimpleDateFormat dateFormat = new SimpleDateFormat("EEEEEE d, MMMMM yyyy\nHH:mm");
         date_in.setText(dateFormat.format(ticket.getDate_in()));
         String dr;
-        if (ticket.getDuration() < 60) dr = ticket.getDuration() + " minutes";
+        if (ticket.getDuration() < 60) dr = (int) ticket.getDuration() + " minutes";
         else
             dr = (int) (ticket.getDuration() / 60.0) + " hours and " + (int) (ticket.getDuration() % 60.0) + " minutes";
         date_check_out.setText(dateFormat.format(ticket.getDate_out()));
         duration.setText(dr);
         String hours;
 
-            hours = ticket.getDuration_paying_for()==1 ? " hour": " hours";
-        duration_paying_for.setText((ticket.getDuration_paying_for())+ hours);
+        hours = ticket.getDuration_paying_for() == 1 ? " hour" : " hours";
+        duration_paying_for.setText((ticket.getDuration_paying_for()) + hours);
         price.setText("£" + ticket.getPrice());
     }
